dockerapps: Start apps inside getCurrent
Users don't normally write docker-compose/docker-app specs that
instruct docker to restart the containers after a reboot. This means
that after applying and update and rebooting, you'll find yourself
with no running containers. This is a bit heavy-handed, but the
call to getCurrent is actually somewhat limited to start-up time-ish
code paths for aktualizr. The idea being, that aktualizr will be
running as a daemon and will therefore get things going for you.

Signed-off-by: Andy Doan <andy@foundries.io>